requirements.txt
Werkzeug==2.2.2
Flask==2.2.2
numpy==1.24.2
onnxruntime==1.19.0
opencv_python==4.10.0.84
Pillow==10.4.0